Add proper counting of pending job arrays in slurm
Pending job arrays are only counted as one job in slurm, since all
information is collected under one job id. Only after the jobs in the
job array are started multiple job ids are assigned. Therefore, this
needs only to be fixed for pending jobs.

The only way I found to count the number of pending jobs in a job array
is to parse the `array_task_str` property of the job array, since slurm
and the pyslurm library do not provide any other information aboutn job
arrays.